// app/routes/public-components/slate-editor/types.d.ts
export type CustomText = {
  bold?: boolean;
  italic?: boolean;
  code?: boolean;
  underline?: boolean;
  strikethrough?: boolean;
  // MARKDOWN PREVIEW SPECIFIC LEAF
  underlined?: boolean;
  title?: boolean;
  list?: boolean;
  hr?: boolean;
  blockquote?: boolean;
  text: string;
  // customs
  indent?: number;
  outdent?: number;
  lineheight?: number;
};

export type CustomTextKey = keyof Omit<CustomText, "text">;